Fix home entry actions
Generated-By: looper 0.8.0 (runner=fixer, agent=codex)
