+++ apps/web/src/components/NewProjectPanel.tsx
@@ -42,39 +42,10 @@ import {
   VIDEO_LENGTHS_SEC,
   VIDEO_MODELS,
 } from '../media/models';
+import { formatPickAndImportFailure } from '../utils/pickAndImportError';
 import { Icon } from './Icon';
 import { Skeleton } from './Loading';
 import { Toast } from './Toast';
-
-/**
- * Best-effort flattening of the `details` field that the
- * pickAndImport main-process handler attaches when the daemon returned
- * a structured error envelope (PR #974 round-4 mrcfps). Daemon errors
- * carry `error.message` and sometimes nested `error.details.reason`;
- * we surface the most operator-actionable string we can find without
- * over-coupling to any particular error code.
- */
-function formatPickAndImportErrorDetails(details: unknown): string | undefined {
-  if (typeof details === 'string' && details.length > 0) return details;
-  if (details == null || typeof details !== 'object') return undefined;
-  const record = details as Record<string, unknown>;
-  const error = record.error;
-  if (error != null && typeof error === 'object') {
-    const errRecord = error as Record<string, unknown>;
-    const message = errRecord.message;
-    const nestedDetails = errRecord.details;
-    if (typeof message === 'string' && message.length > 0) {
-      if (nestedDetails != null && typeof nestedDetails === 'object') {
-        const nestedReason = (nestedDetails as Record<string, unknown>).reason;
-        if (typeof nestedReason === 'string' && nestedReason.length > 0) {
-          return `${message} (${nestedReason})`;
-        }
-      }
-      return message;
-    }
-  }
-  return undefined;
-}
 
 // Snapshot of a curated prompt template, captured at New Project time and
 // folded into ProjectMetadata.promptTemplate. The user may have edited the
@@ -604,16 +575,7 @@ export function NewProjectPanel({
         // network errors). The pickAndImport handler already pre-shapes
         // these into a `{ ok: false, reason, details? }` envelope.
         if ('canceled' in result && result.canceled === true) return;
-        const reason = 'reason' in result && typeof result.reason === 'string'
-          ? result.reason
-          : 'unknown failure';
-        const details = 'details' in result && result.details != null
-          ? formatPickAndImportErrorDetails(result.details)
-          : undefined;
-        setImportFolderError({
-          message: `Open folder failed: ${reason}`,
-          ...(details ? { details } : {}),
-        });
+        setImportFolderError(formatPickAndImportFailure(result));
       } finally {
         setImportingFolder(false);
       }
